+++ modules/non-nixos.nix
@@ -1,8 +1,0 @@
-{
-  hmModule = { nixos, pkgs, lib, ... }: lib.mkIf (!nixos) {
-    home.packages = builtins.attrValues {
-      # Necessary for non-NixOS systems which won't have the flakiest version of Nix
-      inherit (pkgs) nix;
-    };
-  };
-}